Sidebar navigation on smaller screen size
-extract shared code into component and use it in both navigations
-ensure there are no skeleton elements after resizing
