+++ src/components/shared/layouts/Header/Header.module.css
@@ -1,0 +1,4 @@
+.header{
+  padding: var(--chakra-spacing-3);
+  border-bottom: 1px solid var(--chakra-colors-light-800);
+}
